+++ framework/CHANGELOG.md
@@ -4,7 +4,7 @@ Yii Framework 2 Change Log
 2.0.30 under development
 ------------------------
 
-- no changes in this release.
+- Bug #17629: The fallback message formatter is now used if ICU < 4.8 is installed (brandonkelly)
 
 
 2.0.29 October 22, 2019